@@ -49,7 +49,7 @@ dependencies:
   google_mlkit_commons: ^0.7.1
   camera: ^0.11.0+2
   image: ^4.2.0
-  tflite_flutter: ^0.10.4
+  tflite_flutter: ^0.11.0
   sqflite: ^2.3.3+1
   path_provider: ^2.1.3
   image_picker: ^1.1.2
